+++ src/lib/index.ts
@@ -23,6 +23,7 @@ import BeCascader from './be-cascader/BeCascader.svelte';
 import previewImages from './utils/PreviewImage';
 import messageBox from './utils/MessageBox';
 import BeDrawer from './be-drawer/BeDrawer.svelte';
+import { BeCollapse, BeCollapseItem } from './be-collapse';
 
 import { BeTable, BeTableColumn } from './be-table';
 import { BeSelect, BeOption } from './be-select';
@@ -31,4 +32,6 @@ import { BeCheckbox, BeCheckboxGroup } from './be-checkbox';
 import { BeRadio, BeRadioGroup } from './be-radio';
 import { BeMenu, BeMenuItem, BeMenuItemGroup, BeSubMenu } from './be-menu';
 import { BeBreadcrumb, BeBreadcrumbItem } from './be-breadcrumb';
-export { messageBox, BeDrawer, BeCascader, BeRadio, BeRadioGroup, BeButton, previewImages, BeIcon, BeTable, BeTableColumn, BeCarouselItem, BeCarousel, BeAlert, BeDialog, BePagination, BeSwitch, BeTabs, BeTabPane, BeSelect, BeInput, BeDatePicker, BeClock, BeDrag, BeScrollbar, BeProgress, BeUpload, BeNotification, BeLoading, BeCheckbox, BeCheckboxGroup, BeMenu, BeMenuItem, BeMenuItemGroup, BeSubMenu, BeBreadcrumb, BeBreadcrumbItem, BeTree, BeEmpty, BeTimePicker, BeOption };
+import { showNotice, closeALlNotice, closeNotice } from './utils/notice';
+
+export { BeCollapse, BeCollapseItem, showNotice, closeALlNotice, closeNotice, messageBox, BeDrawer, BeCascader, BeRadio, BeRadioGroup, BeButton, previewImages, BeIcon, BeTable, BeTableColumn, BeCarouselItem, BeCarousel, BeAlert, BeDialog, BePagination, BeSwitch, BeTabs, BeTabPane, BeSelect, BeInput, BeDatePicker, BeClock, BeDrag, BeScrollbar, BeProgress, BeUpload, BeNotification, BeLoading, BeCheckbox, BeCheckboxGroup, BeMenu, BeMenuItem, BeMenuItemGroup, BeSubMenu, BeBreadcrumb, BeBreadcrumbItem, BeTree, BeEmpty, BeTimePicker, BeOption };